Adiciona localização de amigos do facebook /me/facebook_friends

+++ accounts/models.py
@@ -7,7 +7,7 @@ from relationships.models import Follow
 
 
 class User(AbstractUser):
-    facebook_id = models.IntegerField(null=True)
+    facebook_id = models.BigIntegerField(null=True)
     about = models.TextField(blank=True, null=True)
     gender = models.CharField(max_length=15, null=True)
     city = models.CharField(max_length=50, null=True)
@@ -21,7 +21,7 @@ class User(AbstractUser):
         through='relationships.Follow',
         related_name='followers',
     )
-    facebook_friends_ids = ArrayField(models.IntegerField(), blank=True, null=True)
+    facebook_friends_ids = ArrayField(models.BigIntegerField(), blank=True, null=True)
 
 
     def follow(self, user):
